Pin the lint rules so CI stops drifting with ruff releases
The workflow installed ruff unpinned with no configuration, so ruff 0.16's
wider default rule set produced 175 errors on a day nobody touched the
repository. Lint runs before the tests, so the 436 tests stopped running
entirely.

ruff.toml lists the rule set explicitly and ignores B008, which is FastAPI's
Depends/Form/Header pattern rather than the mutable-default bug it targets.
The rest are the genuine findings: unused bindings, unsorted imports and
__all__, a deprecated typing import, and two zips that are now strict.

+++ app/modules/matching/service.py
@@ -95,7 +95,7 @@ def score_candidate(job: dict[str, Any], org: dict[str, Any]) -> dict[str, Any]:
     elif job_state and org_state:
         reasons.append(f"different state ({org_state} against {job_state})")
 
-    total = int(round(min(100.0, score)))
+    total = round(min(100.0, score))
     return {
         "score": total,
         "matched_skills": matched,
